requirements.txt
nltk==3.6.1
matplotlib==3.4.2
scipy==1.7.0
flask_ngrok==0.0.25
pandas==1.2.4
requests==2.25.1
Flask==1.1.2
seaborn==0.11.1
scikit_surprise==1.1.1
numpy==1.19.5
IMDbPY==2021.4.18
scikit_learn==0.24.2
surprise==0.1
